Basic setup ready
This commits has the changes on the configuration schema. The client can
define on the transport which realm to look for, as before. But now the
realm can include a `transport-authentication` element. At the moment,
this only includes the SASL configuration.

There is an IT test that won't need more expansion, simple execution
with two nodes. We might add more tests to verify local communication.

The `SASL` protocol implementation doesn't actually exist at this point,
there is only the methods printing to STDOUT.